Bug fixes using Rosenbrock as a test case.
Using Rosenbrock's function as a test case, compared detailed iterative output with an Octave version (qnForDebugging.m). 

The main bug fix was in the linear solve for the search direction. Now I'm using Eigen's llt() function; before I was using something else that was not solving the system correctly (probably it was not the appropriate linear solver).

Now the C++ and Octave versions' iterative output match.